Core Agent Chat Engine
